Implement suggested solution in #1
This solution is not tested, and may be somewhat broken. Throughout the
following week I'm going to be writing a large test-suite for the app,
which will cover whether or not this system works, but not necessarily
if it's fast. Clearly there's going to be an issue if the server crashes
mid background-task. I'll work around that later. I just really wanted
to get this solution in asap to get another prototype up and running.