Simplify deployment process
Include packages that the server depends on in base packages, even if
they aren't used in production. Packages like requests are used for
test client but need to be included in base dependencies to resolve
version conflicts. Separating aerich at that point just creates more
hassle than it solves because the workflow will have to build two
images, one with aerich to use in jobs and one without to use in
production.